requirements.txt
jieba
nltk
python-docx
pyautogui
python-dotenv
sacremoses
sentencepiece
transformers
pandas
datasets
torch
torchvision 
torchaudio
IPython
sacrebleu
evaluate
accelerate
datasets
ctranslate2
hf_hub_ctranslate2
huggingface-hub